# design/pinelabs-openapi.yaml
openapi: 3.0.3
info:
  title: Pine Labs Integration Plugin API
  description: |
    API specification for the Pine Labs Integration Plugin for CARE.

    This plugin allows healthcare facilities to:
    - Configure Pine Labs integration per facility (`PinelabsConfig`), including
      default payment flow and Pine Labs → CARE payment method mappings
      (embedded on the config, similar to eaushadhi's supplier_mappings pattern)
    - Register Pine Labs POS terminals as CARE `emr.Device` records
      (`PinelabsPosTerminal` links a `Device` to a facility's `PinelabsConfig`
      via the dedicated `PATCH /pinelabs_config/{id}/pos-terminals/` endpoint;
      the device's own `client_id`/`store_id` metadata is managed via the core
      CARE `emr/device` API and the `pos-terminal` device-type plugin, not here)
    - Upload payment transactions to a Pine Labs terminal (card/UPI/BharatQR)
    - Poll and refresh transaction status from the Pine Labs Plutus Cloud API
    - Cancel in-flight transactions

    **Care Plugin Conventions:**
    - All `id` fields exposed via API are UUIDs (the table's `external_id`).
      The internal BigInt PK is never exposed.
    - All FK fields exposed via API are UUIDs (the referenced record's `external_id`).
    - `created_by` / `updated_by` are nested user objects on `EMRBaseModel`-backed
      resources (`PinelabsConfig`, `PinelabsPosTerminal`); models built on the
      plain `BaseModel` (`PinelabsPaymentMethodMapping`) only expose
...

...

...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /pinelabs_config/{id}/pos-terminals/:
    get:
      tags:
        - Pinelabs Config
      summary: List POS terminals linked to a facility config
      description: |
        Retrieve all active POS terminals linked to this config — looked up
        by the config's own `id`. Not embedded on
        GET /pinelabs_config/?facility_id= — use this endpoint instead.
      operationId: getPinelabsConfigPosTerminals
      parameters:
        - $ref: '#/components/parameters/ConfigId'
        - name: mine
          in: query
          required: false
          description: |
            If true, only pos terminals for devices belonging to
            organizations the current user has access to are returned
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of active POS terminals for this config
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PinelabsPosTerminal'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags:
        - Pinelabs Config
      summary: Link/unlink POS terminals on a facility config
      description: |
        Full replacement of the POS terminals linked to this config —
        looked up by the config's own `id`. Entries are matched by
        `device_id` (their natural key — a device links to only one pos
        terminal, so no separate `id` is needed): device_ids missing from
        the submitted array are unlinked (soft-deleted), new device_ids
        (existing `emr.Device` records of `care_type=pos-terminal`,
        belonging to this config's facility) are linked.
      operationId: updatePinelabsConfigPosTerminals
      parameters:
        - $ref: '#/components/parameters/ConfigId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PinelabsPosTerminalsUpdate'
            example:
              pos_terminals:
                - device_id: "c2f8d3b5-ae90-4f7c-b401-8e3c9d6f1b25"
                - device_id: "d3a9e4c6-bf01-5f8d-c512-9f4e0a6c2d36"
      responses:
        '200':
          description: Updated list of active POS terminals for this config
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PinelabsPosTerminal'
        '400':
          description: Validation error, e.g. duplicate device_id within the same request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Config, or one of the given device_ids, not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Conflict - one of the given devices is already linked to another Pinelabs POS terminal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                errors:
                  - type: "conflict"
                    msg: "Device POS Terminal 01 is already linked to another Pinelabs POS terminal"
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ---------- Pinelabs Transaction ----------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  parameters:
    ConfigId:
      name: id
      in: path
      required: true
      description: Pinelabs config id (UUID, external_id)
      schema:
        type: string
        format: uuid
  schemas:
    UserSummary:
      type: object
      required:
        - id
        - username
        - first_name
        - last_name
        - email
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email

    # ===================== Payment Method Mapping (embedded) =====================
...
    PinelabsPosTerminalsUpdate:
      type: object
      required:
        - pos_terminals
      properties:
        pos_terminals:
          type: array
          description: |
            Full replacement set of linked POS terminals. See
            PATCH /pinelabs_config/{id}/pos-terminals/ description for
            reconciliation semantics.
          items:
            $ref: '#/components/schemas/PinelabsPosTerminalInput'
...
    DeviceSummary:
      type: object
      description: |
        Trimmed reference to the underlying CARE emr.Device. `client_id` /
        `store_id` come from the device's pos-terminal metadata (managed via
        the core `emr/device` API), not from this plugin.
      properties:
        id:
          type: string
          format: uuid
        registered_name:
          type: string
        care_type:
          type: string
          example: "pos-terminal"
        metadata:
          type: object
          properties:
            client_id:
              type: string
              example: "CLIENT001"
            store_id:
              type: string
              example: "STORE001"
    PinelabsPosTerminalInput:
      type: object
      required:
        - device_id
      properties:
        device_id:
          type: string
          format: uuid
          description: |
            emr.Device id (must be care_type=pos-terminal). A device can be
            linked to only one PinelabsPosTerminal. `device_id` is itself
            the reconciliation key on PATCH — no separate `id` is needed:
            existing links whose device_id is absent from the submitted
            array are soft-deleted; device_ids not yet linked are created.
    PinelabsPosTerminal:
      type: object
      required:
        - id
        - device
        - created_by
        - created_date
        - modified_date
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        device:
          allOf:
            - $ref: '#/components/schemas/DeviceSummary'
          readOnly: true
        created_by:
          allOf:
            - $ref: '#/components/schemas/UserSummary'
          readOnly: true
        updated_by:
          allOf:
            - $ref: '#/components/schemas/UserSummary'
          nullable: true
          readOnly: true
        created_date:
          type: string
          format: date-time
          readOnly: true
          example: "2026-05-19T19:29:35.868682+05:30"
        modified_date:
          type: string
          format: date-time
          readOnly: true
          example: "2026-05-19T19:29:35.868682+05:30"

    # ===================== Pinelabs Transaction =====================
...
    ErrorDetail:
      type: object
      required:
        - type
        - msg
      properties:
        type:
          type: string
          description: Machine-readable error type
          example: "validation_error"
        msg:
          type: string
        code:
          type: integer
          description: Upstream Pine Labs/Plutus response code, present for pinelabs_* error types
    ErrorEnvelope:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
  responses:
...
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            errors:
              - type: "not_authenticated"
                msg: "Authentication credentials were not provided"
    Forbidden:
      description: Forbidden - authenticated but not permitted to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            errors:
              - type: "permission_denied"
                msg: "You do not have permission to perform this action"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            errors:
              - type: "object_not_found"
                msg: "Resource not found"
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            errors:
              - type: "internal_error"
                msg: "Internal server error occurred"